+++ packages/pose-core/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Pose Core adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.2] 2018-04-22
+
+- Adding `initialPose` to `addChild`.
+
+## [0.1.1] 2018-04-22
+
+- Adding lifecycle method for preTransition
+
 ## [0.0.1] 2018-04-18
 
 - First publish